+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ Changelog for the **SODAR Core** Django app package. Loosely follows the
 Note that the issue IDs here refer to ones in the private CUBI GitLab.
 
 
+Unreleased
+==========
+
+Fixed
+-----
+
+- **Projectroles**
+    - Output of template tag ``get_project_link()``
+
+
 v0.6.2 (2019-06-21)
 ===================